Extracted a few networking related items into a new Dagger module and finished the process of mocking the main component for tests.
