+++ app/src/test/kotlin/fr/free/nrw/commons/TestCommonsApplication.kt
@@ -1,9 +1,9 @@
 package fr.free.nrw.commons
 
+import android.content.ContentProviderClient
 import android.content.Context
 import android.content.SharedPreferences
 import android.support.v4.util.LruCache
-import com.google.gson.Gson
 import com.nhaarman.mockito_kotlin.mock
 import com.squareup.leakcanary.RefWatcher
 import fr.free.nrw.commons.auth.AccountUtil
@@ -33,21 +33,31 @@ class TestCommonsApplication : CommonsApplication() {
     override fun setupLeakCanary(): RefWatcher = RefWatcher.DISABLED
 }
 
+@Suppress("MemberVisibilityCanBePrivate")
 class MockCommonsApplicationModule(appContext: Context) : CommonsApplicationModule(appContext) {
     val accountUtil: AccountUtil = mock()
     val appSharedPreferences: SharedPreferences = mock()
     val defaultSharedPreferences: SharedPreferences = mock()
-    val categorySharedPreferences: SharedPreferences = mock()
     val otherSharedPreferences: SharedPreferences = mock()
     val uploadController: UploadController = mock()
     val mockSessionManager: SessionManager = mock()
-    val mediaWikiApi: MediaWikiApi = mock()
     val locationServiceManager: LocationServiceManager = mock()
     val cacheController: CacheController = mock()
     val mockDbOpenHelper: DBOpenHelper = mock()
     val nearbyPlaces: NearbyPlaces = mock()
     val lruCache: LruCache<String, String> = mock()
-    val gson: Gson = Gson()
+    val categoryClient: ContentProviderClient = mock()
+    val contributionClient: ContentProviderClient = mock()
+    val modificationClient: ContentProviderClient = mock()
+    val uploadPrefs: SharedPreferences = mock()
+
+    override fun provideCategoryContentProviderClient(context: Context?): ContentProviderClient = categoryClient
+
+    override fun provideContributionContentProviderClient(context: Context?): ContentProviderClient = contributionClient
+
+    override fun provideModificationContentProviderClient(context: Context?): ContentProviderClient = modificationClient
+
+    override fun providesDirectNearbyUploadPreferences(context: Context?): SharedPreferences = uploadPrefs
 
     override fun providesAccountUtil(context: Context): AccountUtil = accountUtil
 
@@ -60,8 +70,6 @@ class MockCommonsApplicationModule(appContext: Context) : CommonsApplicationModu
     override fun providesUploadController(sessionManager: SessionManager, sharedPreferences: SharedPreferences, context: Context): UploadController = uploadController
 
     override fun providesSessionManager(context: Context, mediaWikiApi: MediaWikiApi, sharedPreferences: SharedPreferences): SessionManager = mockSessionManager
-
-    override fun provideMediaWikiApi(context: Context, sharedPreferences: SharedPreferences, categorySharedPreferences: SharedPreferences, gson: Gson): MediaWikiApi = mediaWikiApi
 
     override fun provideLocationServiceManager(context: Context): LocationServiceManager = locationServiceManager
